+++ src/features/resident/domain/index.ts
@@ -1,0 +1,5 @@
+export * from './entities/Resident';
+export * from './valueObjects/ResidentStatus';
+export * from './valueObjects/Gender';
+export * from './valueObjects/IdentityDocumentType';
+export * from './interfaces/ResidentRepository';